Update Functional
Rewrote table input code to be a template that standardizes the "take inputs, store in a table, offer to do it again" activity. Need to go back through existing files that have mismatched implementations (different sessions with Copilot introduced variances) and redo some of them that don't need so much customization.